+++ tests/formatter/input-output-pairs/multiline_properties.in.gd
@@ -23,6 +23,8 @@ var p8:
 var p9:
 	set(_x): pass
 	get: return 1
+var p9x:
+	get(): return 1
 var p10:
 	get:
 		var xyz = 123